+++ docs/source/offline.rst
@@ -1,0 +1,23 @@
+Offline Estimators (``m.estimators.offline``)
+=============================================
+
+.. automodule:: moirae.estimators.offline
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Scipy Estimators (``m.e.offline.scipy``)
+----------------------------------------
+
+.. automodule:: moirae.estimators.offline.scipy
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Loss Functions (``m.e.offline.loss``)
+-------------------------------------
+
+.. automodule:: moirae.estimators.offline.loss
+   :members:
+   :undoc-members:
+   :show-inheritance:
